Is this... better?
I have this intuition that an actor system is a neato place to start for
a chatbot, but I'm not sure how it's going to grow as commands and other
weirdo interactions are added. Maybe instead of guessing what those
weirdo interactions could be I'll wait until they show up and refactor.
But gosh darnit, spending a lot of time architecting the unknown future
is so fun. Kind of.